modularized App.js RootStackScreen
